fix: correct some logic and UI bugs, add assets for readme.md

+++ Client/publisher/publisherbookwidget.h
@@ -41,7 +41,7 @@ private slots:
     int m_publisherID;
 
     int m_bookID;
-    int m_discount;
+    double m_discount;
 
     void loadBooks();
     void handleGetBooks(const QJsonObject& data);
